move countDigits to separate file
